+++ nost_tools/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "2.4.0"
+__version__ = "3.0.0"
 
 from .application import Application
 from .application_utils import ConnectionConfig, ModeStatusObserver, TimeStatusPublisher
@@ -6,7 +6,7 @@ from .configuration import ConnectionConfig
 from .entity import Entity
 from .logger_application import LoggerApplication
 from .managed_application import ManagedApplication
-from .manager import Manager, TimeScaleUpdate
+from .manager import Manager
 from .observer import Observable, Observer
 from .publisher import ScenarioTimeIntervalPublisher, WallclockTimeIntervalPublisher
 from .schemas import (